VCPkg support
- Removed all included dependencies.
- Configured project to use dependencies from VCPkg.
- Cleaned-up include and library paths.
- Included 'Setup.h' file for wxWidgets.
- Updated build instructions.

+++ KxFramework/HTML/serialize.cpp
@@ -229,7 +229,7 @@ static std::string serialize(GumboNode* node) {
 
   // build attr string  
   const GumboVector * attribs = &node->v.element.attributes;
-  for (int i=0; i< attribs->length; ++i) {
+  for (size_t i = 0; i < static_cast<size_t>(attribs->length); ++i) {
     GumboAttribute* at = static_cast<GumboAttribute*>(attribs->data[i]);
     atts.append(build_attributes(at, no_entity_substitution));
   }
